requirements.txt
fastapi
uvicorn
python-multipart
requests
pandas
matplotlib
seaborn
openpyxl
autogen
aiofiles
jinja2
sqlalchemy
openai
python-dotenv